refactor: 사용자 입력받는 기능을 InputView로 합침, 입력값 검증 및 재입력기능 제거
