+++ src/main/java/lotto/LottoGameApplication.java
@@ -2,12 +2,12 @@ package lotto;
 
 import lotto.controller.LottoGameController;
 import lotto.domain.LottoRandomNumbersGenerator;
-import lotto.service.InputService;
+import lotto.view.InputView;
 
 public class LottoGameApplication {
 
 	public static void main(String[] args) {
-		LottoGameController controller = new LottoGameController(new InputService(), new LottoRandomNumbersGenerator());
+		LottoGameController controller = new LottoGameController(new InputView(), new LottoRandomNumbersGenerator());
 		controller.play();
 	}
