+++ src/main/java/lotto/controller/LottoGameController.java
@@ -4,7 +4,6 @@ import lotto.domain.LottoGame;
 import lotto.domain.LottoNumbers;
 import lotto.domain.LottoResult;
 import lotto.domain.RandomNumbersGenerator;
-import lotto.service.InputService;
 import lotto.view.InputView;
 import lotto.view.ResultView;
 
@@ -12,24 +11,24 @@ import java.util.List;
 
 public class LottoGameController {
 
-	private InputService lottoInputService;
+	private InputView inputView;
 	private RandomNumbersGenerator randomNumbersGenerator;
 
-	public LottoGameController(InputService lottoInputService, RandomNumbersGenerator randomNumbersGenerator) {
-		this.lottoInputService = lottoInputService;
+	public LottoGameController(InputView inputView, RandomNumbersGenerator randomNumbersGenerator) {
+		this.inputView = inputView;
 		this.randomNumbersGenerator = randomNumbersGenerator;
 	}
 
 	public void play() {
 
 		//로또구입
-		int purchaseAmount = this.lottoInputService.inputPurchaseAmount();
+		int purchaseAmount = this.inputView.inputPurchaseAmount();
 		LottoGame lottoGame = new LottoGame();
 		List<LottoNumbers> lottos = lottoGame.purchaseLottos(purchaseAmount, this.randomNumbersGenerator);
 		InputView.printPurchaseLottos(lottos);
 
 		//당첨확인
-		List<Integer> winLottoNumbers = this.lottoInputService.inputWinLottoNumbers();
+		List<Integer> winLottoNumbers = this.inputView.inputWinLottoNumbers();
 		LottoResult lottoResult = new LottoResult(lottos, new LottoNumbers(winLottoNumbers));
 		ResultView.printLottoResult(lottoResult);
